Resolved issue where final buffer wasn't getting fully flashed. Added check to ensure buff1 is done before quitting and resetting buffers. Currently have a hacked load_oper_info_elements_chr function to test CHR-ROM flashing. NROM PRG & CHR flashing and dumping tested and working on black ASUS win10 machine. Firmware still doesn't implement previously added feature for oper_info rd/wr_func pointer.  flash/dump.c simply overrides by providing the function pointer based on memory type.
A bit frustrated with how complicated everything is currently...
Feel like there's a lack of standard for function definition and calls.
Would like majority of host program flow to be determined solely by
script calls.  Firmware is getting more generic and versatile at the cost
of being SLOW!!!  Perhaps that's okay though as it's not terribly slow in
comparision to comparable AVR kazzo firmware builds.  And with current
target to migrate to STM32 device this generic setup is much cleaner and
easier to work with.  However complexity was added to the firmware code in
order to help make it faster.  Not sure the goal of faster should take
priority over speed.  Obsessing over speed will cause issues with race
conditions, and complicate code to the point where I won't reach my goal.

With the long break since I've been working on this project it's making me
more willing to scrap some previous work for a cleaner solution more
aligned with my end goal.  I had to write some of this code to get a
better idea of my what my end goal will look like anyway.  Going to make a
couple builds for UIROM, and action53 as they're urgently needed.  But
beyond that I need to start working on script features, and STM32 support!